refactor: restructure Shelf domain model and decouple from Book module using ports
Context:
This change improves the architectural boundaries within the Library domain by
reorganizing the Shelf entity into a proper package structure and eliminating
direct dependencies between the Shelf and Book bounded contexts. Previously,
ShelfDomainRepositoryImpl had a direct dependency on BookDomainRepository,
violating the Dependency Inversion Principle and creating tight coupling
between modules.

What changed:

Domain Layer:
• Moved Shelf.java from core.domain/ to core.domain.model/ package for better organization
• Created BookAccessPort interface in shelf/core/domain/ports/ (outbound port)
  - Defines what Shelf module needs from Book context in its own terms
  - Follows hexagonal architecture's port-adapter pattern

Infrastructure Layer:
• Updated ShelfDomainRepositoryImpl to depend on BookAccessPort instead of BookDomainRepository
• Removed duplicate ShelfJpaRepository field (kept single jpaRepository reference)
• Simplified book ID retrieval logic using the new port

Application/Mapping Layer:
• Updated all import statements across ShelfService, ShelfDomainRepository, and ShelfMapper
  to reflect the new Shelf location (core.domain.model.Shelf)

Cataloging (Book) Module:
• Created BookAccessAdapter in book/infrastructure/adapter/
  - Implements Shelf's BookAccessPort interface
  - Adapts Book domain's BookDomainRepository to Shelf's needs
  - Dependency arrow points FROM Book TO Shelf's port (DIP compliance)

Notes/Risks:
• This is a structural refactoring with no functional changes to business logic
• The port-adapter pattern enables independent evolution of Shelf and Book modules
• Future work: Consider similar decoupling for other cross-context dependencies
• No database migrations required; changes are code-level only
