+++ src/main/java/com/penrose/bibby/library/stacks/shelf/core/domain/ShelfDomainRepository.java
@@ -1,5 +1,6 @@
 package com.penrose.bibby.library.stacks.shelf.core.domain;
 
+import com.penrose.bibby.library.stacks.shelf.core.domain.model.Shelf;
 import com.penrose.bibby.library.stacks.shelf.core.domain.valueobject.ShelfId;
 
 public interface ShelfDomainRepository {
