+++ src/main/java/com/penrose/bibby/library/stacks/shelf/core/domain/model/Shelf.java
@@ -1,4 +1,4 @@
-package com.penrose.bibby.library.stacks.shelf.core.domain;
+package com.penrose.bibby.library.stacks.shelf.core.domain.model;
 
 import com.penrose.bibby.library.stacks.shelf.core.domain.valueobject.ShelfId;
 import java.util.List;
