+++ src/main/java/com/penrose/bibby/library/stacks/shelf/infrastructure/adapter/outbound/ShelfDomainRepositoryImpl.java
@@ -1,14 +1,12 @@
 package com.penrose.bibby.library.stacks.shelf.infrastructure.adapter.outbound;
 
-import com.penrose.bibby.library.cataloging.book.core.domain.Book;
-import com.penrose.bibby.library.cataloging.book.core.domain.BookDomainRepository;
-import com.penrose.bibby.library.stacks.shelf.core.domain.Shelf;
 import com.penrose.bibby.library.stacks.shelf.core.domain.ShelfDomainRepository;
+import com.penrose.bibby.library.stacks.shelf.core.domain.model.Shelf;
+import com.penrose.bibby.library.stacks.shelf.core.domain.ports.BookAccessPort;
 import com.penrose.bibby.library.stacks.shelf.core.domain.valueobject.ShelfId;
 import com.penrose.bibby.library.stacks.shelf.infrastructure.entity.ShelfEntity;
 import com.penrose.bibby.library.stacks.shelf.infrastructure.mapping.ShelfMapper;
 import com.penrose.bibby.library.stacks.shelf.infrastructure.repository.ShelfJpaRepository;
-import java.util.ArrayList;
 import java.util.List;
 import org.springframework.stereotype.Component;
 
@@ -17,35 +15,26 @@ public class ShelfDomainRepositoryImpl implements ShelfDomainRepository {
 
   private final ShelfJpaRepository jpaRepository;
   private final ShelfMapper shelfMapper;
-  private final BookDomainRepository bookDomainRepository;
-  private final ShelfJpaRepository shelfJpaRepository;
+  private final BookAccessPort bookAccessPort;
 
   public ShelfDomainRepositoryImpl(
-      ShelfJpaRepository jpaRepository,
-      ShelfMapper shelfMapper,
-      BookDomainRepository bookDomainRepository,
-      ShelfJpaRepository shelfJpaRepository) {
+      ShelfJpaRepository jpaRepository, ShelfMapper shelfMapper, BookAccessPort bookAccessPort) {
     this.jpaRepository = jpaRepository;
     this.shelfMapper = shelfMapper;
-    this.bookDomainRepository = bookDomainRepository;
-    this.shelfJpaRepository = shelfJpaRepository;
+    this.bookAccessPort = bookAccessPort;
   }
 
   @Override
   public Shelf getById(ShelfId id) {
     ShelfEntity entity = jpaRepository.findById(id.shelfId()).orElse(null);
-    List<Book> books = bookDomainRepository.getBooksByShelfId(id.shelfId());
-    List<Long> bookIds = new ArrayList<>();
-    for (Book book : books) {
-      bookIds.add(book.getBookId().getId());
-    }
+    List<Long> bookIds = bookAccessPort.getBookIdsByShelfId(id.shelfId());
     return shelfMapper.toDomainFromDTO(entity, bookIds);
   }
 
   @Override
   public void save(Shelf shelf) {
-    ShelfEntity entity = shelfJpaRepository.findById(shelf.getId()).get();
+    ShelfEntity entity = jpaRepository.findById(shelf.getId()).get();
     //        entity = shelfMapper.updateEntity(shelf);
-    shelfJpaRepository.save(entity);
+    jpaRepository.save(entity);
   }
 }
